[feature] Move lists actions to client side

+++ app/api/lists/route.ts
@@ -1,4 +1,4 @@
-import { ListAction, ListType } from "@/lib/enums";
+import { ListAction, ListState, ListType } from "@/lib/enums";
 import {
     HttpResponse,
     MissingBodyError
@@ -6,14 +6,19 @@ import {
 import { VercelPoolClient, db } from "@vercel/postgres"
 import { NextRequest } from "next/server"
 
-export const getListsByUser = async (userId: string, gameId: number): Promise<any[]> => {
-    const client = await db.connect()
+export const get = async (userId: string, gameId: number, client: VercelPoolClient): Promise<Response> => {
+    if (!userId || !gameId) return HttpResponse(null, false, MissingBodyError)
     const { rows } = await client.query(
         `SELECT * FROM list_item WHERE user_id = $1 AND game = $2;`,
         [userId, gameId]
     )
 
-    return rows
+    return HttpResponse({
+        [ListType.Favorite]: rows.some((list) => list.list_type === ListType.Favorite) ? ListState.Active : ListState.Inactive,
+        [ListType.Playlist]: rows.some((list) => list.list_type === ListType.Playlist) ? ListState.Active : ListState.Inactive,
+        [ListType.Finished]: rows.some((list) => list.list_type === ListType.Finished) ? ListState.Active : ListState.Inactive,
+        [ListType.Custom]: ListState.Inactive
+    }, true)
 }
 
 const add = async (
@@ -23,6 +28,7 @@ const add = async (
     listType: ListType,
     client: VercelPoolClient
 ): Promise<Response> => {
+    if (!userId || !gameId || !listType) return HttpResponse(null, false, MissingBodyError)
     const { rowCount } = await client.query('INSERT INTO list_item VALUES ($1, $2, $3, $4);', [
         userId,
         listType,
@@ -45,6 +51,7 @@ const remove = async (
     listType: ListType,
     client: VercelPoolClient
 ): Promise<Response> => {
+    if (!userId || !gameId || !listType) return HttpResponse(null, false, MissingBodyError)
     const { rowCount } = await client.query(`DELETE FROM list_item WHERE
         user_id = $1 AND
         list_type = $2 AND
@@ -63,11 +70,12 @@ const remove = async (
 export async function POST(_request: NextRequest) {
     const client = await db.connect()
     const { userId, gameId, listId, listType, action } = await _request.json()
-    if (!userId || !gameId || !listType || !action) return HttpResponse(null, false, MissingBodyError)
+    if (!action) return HttpResponse(null, false, MissingBodyError)
 
     const actions: Record<ListAction, () => Promise<Response>> = {
         [ListAction.Add]: () => add(userId, gameId, listId, listType, client),
-        [ListAction.Remove]: () => remove(userId, gameId, listId, listType, client)
+        [ListAction.Remove]: () => remove(userId, gameId, listId, listType, client),
+        [ListAction.Get]: () => get(userId, gameId, client)
     }
 
     try {
